# openapi.yaml
openapi: 3.0.0
info:
  title: SysAcademy API
  description: API para gerenciamento de sistema acadêmico
  version: 1.0.0
  contact:
    name: Gustavo Cruz
    url: https://github.com/gustavvocruzz
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT

servers:
  - url: http://localhost:8080
    description: Local Development Server

tags:
  - name: Alunos
    description: Endpoints para gerenciamento de alunos

paths:
  /alunos:
    get:
      summary: Listar alunos
      description: Retorna uma lista paginada de alunos com filtros opcionais
      operationId: listarAlunos
      tags:
        - Alunos
      parameters:
        - name: nome
          in: query
          description: Filtro por nome do aluno (busca parcial)
          schema:
            type: string
          example: João
        - name: email
          in: query
          description: Filtro por e-mail do aluno (busca parcial)
          schema:
            type: string
          example: joao@example.com
        - name: celular
          in: query
          description: Filtro por celular do aluno (busca parcial)
          schema:
            type: string
          example: "11987654321"
        - name: cidade
          in: query
          description: Filtro por cidade do aluno
          schema:
            type: string
          example: São Paulo
        - name: estado
          in: query
          description: Filtro por estado (UF) do aluno
          schema:
            type: string
          example: SP
        - name: page
          in: query
          description: Número da página (começa em 0)
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          description: Quantidade de registros por página
          schema:
            type: integer
            default: 20
        - name: sort
          in: query
          description: "Campo para ordenação (ex: nome,asc ou dataCadastro,desc)"
          schema:
            type: string
      responses:
        "200":
          description: Lista de alunos retornada com sucesso
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageAlunoResponse"
        "400":
          description: Parâmetros de paginação inválidos

    post:
      summary: Cadastrar novo aluno
      description: Cria um novo aluno no sistema
      operationId: cadastrarAluno
      tags:
        - Alunos
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AlunoRequest"
      responses:
        "201":
          description: Aluno cadastrado com sucesso
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlunoResponse"
        "400":
          description: Dados inválidos fornecidos

  /alunos/{id}:
    get:
      summary: Buscar aluno por ID
      description: Retorna os dados de um aluno específico
      operationId: buscarAlunoPorId
      tags:
        - Alunos
      parameters:
        - name: id
          in: path
          required: true
          description: ID do aluno
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Aluno encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlunoResponse"
        "404":
          description: Aluno não encontrado

    put:
      summary: Atualizar aluno
      description: Atualiza os dados de um aluno existente
      operationId: atualizarAluno
      tags:
        - Alunos
      parameters:
        - name: id
          in: path
          required: true
          description: ID do aluno
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AlunoRequest"
      responses:
        "200":
          description: Aluno atualizado com sucesso
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlunoResponse"
        "404":
          description: Aluno não encontrado
        "400":
          description: Dados inválidos fornecidos

    delete:
      summary: Deletar aluno
      description: Remove um aluno do sistema
      operationId: deletarAluno
      tags:
        - Alunos
      parameters:
        - name: id
          in: path
          required: true
          description: ID do aluno
          schema:
            type: integer
            format: int64
      responses:
        "204":
          description: Aluno deletado com sucesso
        "404":
          description: Aluno não encontrado

components:
  schemas:
    AlunoRequest:
      type: object
      description: Dados para criar ou atualizar um aluno
      required:
        - nome
      properties:
        nome:
          type: string
          maxLength: 150
          description: Nome completo do aluno
          example: João da Silva
        dataNascimento:
          type: string
          format: date
          description: Data de nascimento do aluno
          example: "2000-05-15"
        genero:
          type: string
          maxLength: 1
          description: Gênero do aluno (M ou F)
          example: M
        telefone:
          type: string
          maxLength: 30
          description: Telefone residencial
          example: "(11) 3456-7890"
        celular:
          type: string
          maxLength: 30
          description: Número de celular
          example: "(11) 98765-4321"
        email:
          type: string
          format: email
          maxLength: 150
          description: Endereço de e-mail
          example: joao@example.com
        observacao:
          type: string
          description: Observações adicionais sobre o aluno
        endereco:
          type: string
          maxLength: 150
          description: Endereço residencial
          example: "Rua Principal, 123"
        complemento:
          type: string
          maxLength: 150
          description: Complemento do endereço
          example: "Apto 42"
        bairro:
          type: string
          maxLength: 100
          description: Bairro
          example: Centro
        cidade:
          type: string
          maxLength: 100
          description: Cidade
          example: São Paulo
        estado:
          type: string
          maxLength: 2
          description: Unidade Federativa (UF)
          example: SP
        cep:
          type: string
          maxLength: 25
          description: Código de Endereçamento Postal
          example: "01310-100"

    AlunoResponse:
      type: object
      description: Resposta com os dados do aluno
      properties:
        id:
          type: integer
          format: int64
          description: ID único do aluno
          example: 1
        nome:
          type: string
          description: Nome completo do aluno
          example: João da Silva
        dataNascimento:
          type: string
          format: date
          description: Data de nascimento do aluno
          example: "2000-05-15"
        genero:
          type: string
          description: Gênero do aluno
          example: M
        celular:
          type: string
          description: Número de celular
          example: "(11) 98765-4321"
        email:
          type: string
          description: Endereço de e-mail
          example: joao@example.com
        cidade:
          type: string
          description: Cidade do aluno
          example: São Paulo
        estado:
          type: string
          description: Unidade Federativa (UF)
          example: SP
        dataCadastro:
          type: string
          format: date-time
          description: Data e hora do cadastro
          example: "2024-01-15T10:30:00"

    PageAlunoResponse:
      type: object
      description: Página de alunos
      properties:
        content:
          type: array
          description: Lista de alunos
          items:
            $ref: "#/components/schemas/AlunoResponse"
        pageable:
          type: object
          description: Informações de paginação
          properties:
            pageNumber:
              type: integer
            pageSize:
              type: integer
            sort:
              type: object
              properties:
                empty:
                  type: boolean
                sorted:
                  type: boolean
                unsorted:
                  type: boolean
            offset:
              type: integer
            paged:
              type: boolean
            unpaged:
              type: boolean
        totalElements:
          type: integer
          description: Total de elementos
        totalPages:
          type: integer
          description: Total de páginas
        last:
          type: boolean
          description: Se é a última página
        size:
          type: integer
          description: Tamanho da página
        number:
          type: integer
          description: Número da página atual
        sort:
          type: object
          description: Informações de ordenação
        numberOfElements:
          type: integer
          description: Número de elementos nesta página
        first:
          type: boolean
          description: Se é a primeira página
        empty:
          type: boolean
          description: Se a página está vazia